modularize floating toolbar
